- enhanced DidYouMean computation using a faster count on index entries; this causes that results can be ranked better - added limitations on DidYouMean result sets according to input and output string length
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@7246 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/net/yacy/kelondro/index/RowCollection.java
@@ -121,7 +121,7 @@ public class RowCollection implements Iterable<Row.Entry>, Cloneable {
     }
     
     protected RowCollection(final Row rowdef, final Row.Entry exportedCollectionRowEnvironment) {
-        final int chunkcachelength = exportedCollectionRowEnvironment.cellwidth(1) - exportOverheadSize;
+        final int chunkcachelength = exportedCollectionRowEnvironment.cellwidth(1) - (int) exportOverheadSize;
         final Row.Entry exportedCollection = exportRow(chunkcachelength).newEntry(exportedCollectionRowEnvironment, 1);
         
         this.rowdef = rowdef;
@@ -192,7 +192,7 @@ public class RowCollection implements Iterable<Row.Entry>, Cloneable {
     
     private static Column exportColumn0, exportColumn1, exportColumn2, exportColumn3, exportColumn4;
 
-    protected static final int exportOverheadSize = 14;
+    protected static final long exportOverheadSize = 14;
 
     private static Row exportRow(final int chunkcachelength) {
         /*
